# project-dam-backend/dist/dist/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
            "mongoose.Types.ObjectId": {
                "type": "string"
            },
...
            "IDevice": {
                "properties": {
                    "vendor": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "serialNumber": {
                        "type": "string"
                    },
                    "macAddress": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "site": {
                        "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                    },
                    "connectedDevices": {
                        "items": {
                            "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                        },
                        "type": "array"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "vendor",
                    "category",
                    "type",
                    "serialNumber",
                    "macAddress",
                    "state",
                    "site",
                    "connectedDevices",
                    "isActive"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "CreateDeviceRequest": {
                "properties": {
                    "vendor": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "serialNumber": {
                        "type": "string"
                    },
                    "macAddress": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "site": {
                        "type": "string"
                    },
                    "connectedDevices": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "vendor",
                    "category",
                    "type",
                    "serialNumber",
                    "macAddress",
                    "state",
                    "site"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UpdateDeviceRequest": {
                "properties": {
                    "vendor": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "serialNumber": {
                        "type": "string"
                    },
                    "macAddress": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "site": {
                        "type": "string"
                    },
                    "connectedDevices": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "ICable": {
                "properties": {
                    "device1": {
                        "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                    },
                    "device2": {
                        "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                    },
                    "cableType": {
                        "type": "string"
                    },
                    "interface1": {
                        "properties": {
                            "duplex": {
                                "type": "string"
                            },
                            "linkSpeed": {
                                "type": "string"
                            },
                            "portId": {
                                "type": "string"
                            },
                            "device": {
                                "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                            }
                        },
                        "required": [
                            "duplex",
                            "linkSpeed",
                            "portId",
                            "device"
                        ],
                        "type": "object"
                    },
                    "interface2": {
                        "properties": {
                            "duplex": {
                                "type": "string"
                            },
                            "linkSpeed": {
                                "type": "string"
                            },
                            "portId": {
                                "type": "string"
                            },
                            "device": {
                                "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                            }
                        },
                        "required": [
                            "duplex",
                            "linkSpeed",
                            "portId",
                            "device"
                        ],
                        "type": "object"
                    }
                },
                "required": [
                    "device1",
                    "device2",
                    "cableType",
                    "interface1",
                    "interface2"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "CreateCableRequest": {
                "properties": {
                    "device1": {
                        "type": "string"
                    },
                    "device2": {
                        "type": "string"
                    },
                    "cableType": {
                        "type": "string"
                    },
                    "interface1": {
                        "properties": {
                            "duplex": {
                                "type": "string"
                            },
                            "linkSpeed": {
                                "type": "string"
                            },
                            "portId": {
                                "type": "string"
                            },
                            "device": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "duplex",
                            "linkSpeed",
                            "portId",
                            "device"
                        ],
                        "type": "object"
                    },
                    "interface2": {
                        "properties": {
                            "duplex": {
                                "type": "string"
                            },
                            "linkSpeed": {
                                "type": "string"
                            },
                            "portId": {
                                "type": "string"
                            },
                            "device": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "duplex",
                            "linkSpeed",
                            "portId",
                            "device"
                        ],
                        "type": "object"
                    }
                },
                "required": [
                    "device1",
                    "device2",
                    "cableType",
                    "interface1",
                    "interface2"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UpdateCableRequest": {
                "properties": {
                    "device1": {
                        "type": "string"
                    },
                    "device2": {
                        "type": "string"
                    },
                    "cableType": {
                        "type": "string"
                    },
                    "interface1": {
                        "properties": {
                            "duplex": {
                                "type": "string"
                            },
                            "linkSpeed": {
                                "type": "string"
                            },
                            "portId": {
                                "type": "string"
                            },
                            "device": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "interface2": {
                        "properties": {
                            "duplex": {
                                "type": "string"
                            },
                            "linkSpeed": {
                                "type": "string"
                            },
                            "portId": {
                                "type": "string"
                            },
                            "device": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "project-dam-backend",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "paths": {
...
        "/devices": {
            "get": {
                "operationId": "GetDevices",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pages": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "devices": {
                                            "items": {
                                                "$ref": "#/components/schemas/IDevice"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": [
                                        "pages",
                                        "total",
                                        "devices"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Devices"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "default": 1,
                            "format": "double",
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "required": false,
                        "schema": {
                            "default": 10,
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "post": {
                "operationId": "CreateDevice",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IDevice"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Devices"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDeviceRequest"
                            }
                        }
                    }
                }
            }
        },
        "/devices/{deviceId}": {
            "get": {
                "operationId": "GetDevice",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IDevice"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Devices"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "deviceId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateDevice",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IDevice"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Devices"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "deviceId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDeviceRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteDevice",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Devices"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "deviceId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/cables": {
            "get": {
                "operationId": "GetCables",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pages": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "cables": {
                                            "items": {
                                                "$ref": "#/components/schemas/ICable"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": [
                                        "pages",
                                        "total",
                                        "cables"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cables"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "default": 1,
                            "format": "double",
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "required": false,
                        "schema": {
                            "default": 10,
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "post": {
                "operationId": "CreateCable",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ICable"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cables"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCableRequest"
                            }
                        }
                    }
                }
            }
        },
        "/cables/{cableId}": {
            "get": {
                "operationId": "GetCable",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ICable"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cables"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cableId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateCable",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ICable"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cables"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cableId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCableRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteCable",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cables"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cableId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}